drive_watcher/requirements.txt
google-api-python-client==2.133.0
google-auth==2.30.0
google-cloud-secret-manager==2.20.0